Use web workers to run pyodide code away from the ui thread
