+++ game_frontend/djangoBundler.js
@@ -9,16 +9,14 @@ const outDir = Path.join(__dirname, '../aimmo/static/react')
 const options = {
   outDir,
   outFile: 'index.html',
-  publicUrl: './',
+  publicUrl: '/static/react/',
   watch: process.env.NODE_ENV !== 'production',
   minify: process.env.NODE_ENV === 'production',
   target: 'browser',
   cache: process.env.NODE_ENV === 'production'
 }
 
-const templateFolder = Path.resolve(
-  Path.join(__dirname, '../aimmo/templates/players')
-)
+const templateFolder = Path.resolve(Path.join(__dirname, '../aimmo/templates/players'))
 const handlebarsTemplatePath = Path.resolve(
   Path.join(__dirname, './public/handlebars_template.html')
 )
@@ -26,7 +24,7 @@ const handlebarsTemplatePath = Path.resolve(
 const bundler = new Bundler(file, options)
 
 function getReactURL (entryPointHTML) {
-  const regex = /(<script src=")(.*\.js)("><\/script>)/g
+  const regex = /(<script src="\/static\/react\/)(.*\.js)("><\/script>)/g
   return regex.exec(entryPointHTML)[2]
 }
