+++ game_frontend/package.json
@@ -54,7 +54,6 @@
     "ace-builds": "^1.4.11",
     "babylon": "^6.18.0",
     "babylonjs": "^4.1.0",
-    "comlink": "^4.3.0",
     "core-js": "3",
     "eslint-plugin-cypress": "^2.10.3",
     "handlebars": "^4.7.6",
@@ -73,6 +72,7 @@
     "rxjs": "^6.5.2",
     "socket.io-client": "^2.1.1",
     "styled-components": "^3.4.10",
+    "threads": "^1.6.3",
     "webfontloader": "^1.6.28"
   },
   "esLintConfig": {
@@ -82,6 +82,11 @@
       }
     }
   },
+  "browserslist": [
+    "last 3 chrome versions",
+    "last 3 firefox versions",
+    "last 3 safari versions"
+  ],
   "scripts": {
     "lint": "eslint 'src/**/*.{js,ts}'",
     "test": "jest"
