+++ game_frontend/public/handlebars_template.html
@@ -80,6 +80,5 @@
   {{/if_eq}}
   <!-- End Google Tag Manager (noscript) -->
   <div id="root"></div>
-  <script src="https://pyodide-cdn2.iodide.io/v0.15.0/full/pyodide.js"></script>
   <script src="{% static '{{ reactUrl }}' %}"></script>
 </body>
