+++ game_frontend/public/index.html
@@ -2,3 +2,4 @@
 <title>Kurono</title>
 <div id="root"></div>
 <script src="../src/index.js"></script>
+<script src="../src/pyodide/webWorker.ts"></script>
