+++ game_frontend/src/index.js
@@ -1,5 +1,6 @@
 import 'core-js/stable'
 import 'regenerator-runtime/runtime'
+import 'threads/register'
 import React from 'react'
 import { render } from 'react-dom'
