+++ game_frontend/src/pyodide/computedTurnResult.ts
@@ -1,0 +1,5 @@
+export default interface ComputedTurnResult {
+  action: object
+  log: string
+  turnCount: number
+}
